refactor(aiaccountstatus): split unbound helpers out of service.go
service.go sat at the file-size ratchet's baseline (990 lines); the xAI
quota-attribution fix's exhausted-threshold check pushed it to 996 and
tripped the "may only shrink" gate. flightKey, reconcileTenantBindings,
sanitizeMsg, firstNonEmpty and metadataString take no Service receiver
and don't belong in this file over any other in the package, so they
move to service_helpers.go rather than earning the growth an exception.
